Bug Fixes. Location and search view updates.
Updates:
-add cancel button on select location view
-add title "add location" after selecting location
-search bar remove cancel button, remove searched annotations when textfield is cleared.
-add “added date” for saved locations
-select locations view title changed to "near dropped pin or current location"
Fixes:
-deleted alert is sent to all users
-cost textfield done not working
-textfield above keyboard
-searched locations directions left button

+++ Airvolution/User.m
@@ -43,7 +43,7 @@
     return [NSSet setWithArray: array];
 }
 -(NSString *)points{
-    return [NSString stringWithFormat:@"%lu", self.locations.count];
+    return [NSString stringWithFormat:@"%lu", (unsigned long)self.locations.count];
 }
